refactor(agent): consolidate conversation domain to single index module
